enable v3 app ssh via UI for web: type process
currently the UI fails establishing an ssh connection to apps that rely on cc v3
functionality. Some basic apps ( e.g. `cf push test -b staticfile_buildpack` seems
to produce an app that the UI can still ssh into ) still work but anything that
relies on v3 functionality ( e.g. multi process apps or `cf push --strategy rolling` )
will fail.

The default type for a process of a pushed app is `web`. This commit enables using the UI
to ssh to the default web process of V3 apps. It left out adding support for all other process
types since the UI misses the feature to select nested processes. It only displays the `instances`
that an app has running but none of the nested elements of V3 apps.

we only want to ignore 404 errors when checking for v3 availability. other errors may still provide
relevant information.

some auth related code was extracted from getSSHCode so it's reusable in CheckForV3... in the same
reasoning requesting the token and refreshing it was moved up.